Port CW3 to Secret
This commit contains a large refactor to the internals of the
cw3-fixed-multisig implementation to work on Secret and without
cw-storage-plus as a dependency.
It includes migrating from the usual Map to a similar Keymap,
as well as introducing an implementation of a Binary Search Tree
(based on the secret-toolkit Item implementation) to maintain
ordering in state without massive, expensive sorting operations
at any one time.

Signed-off-by: Bobby <bobby@apollo.farm>